Refactor: Added a generic CapstoneDisassembler class.
The "Disassembler" class has been refactored to make it more generic; to
implement a new Capstone-supported architecture you just have to inherit
from it and provide the missing (pure virtual) methods that handle the
arch-specific details needed by the base class to produce output
suitable for remill. Hooks are also present, so that the inheriting
class is able to adjust (or entirely rewrite) the computed data before
it is passed to the lifter.