Refactor, fixed a bug in the PostDecodeHook (see details).
remill::Instruction's constructor is now public; there were no
internal resources that needed protection, and it comes in useful
for testdec to actually access the same data structure that remill
would use when lifting. Another alternative could be to make a
friend function specifically for testdec.

The PostDecodeHook is used to adjust the remill::Instruction object
that has been generated by the CapstoneDisassembler class. The hook
was being called after the semantic function name was generated, thus
ignoring the changes made by inside the callback.